requirements.txt
PyMuPDF==1.24.5
pyudev==0.24.3
pyopenssl==24.1.0
